Refactor neurosymbolic pipeline into modular package
