+++ pipeline/__init__.py
@@ -1,0 +1,8 @@
+"""Neurosymbolic pipeline package with modular preprocessing, reasoning, and evaluation stages."""
+
+__all__ = [
+    "preprocess",
+    "symbolic",
+    "evaluation",
+    "run_pipeline",
+]
